Fix formatting and review content
I fixed a number of inconsistencies in the way that long and short
hyphens were used through the document. I also simplified some of the
text for older and less relevant jobs to make the wording more concise
and easier to read, and used bold formatting for the bullet 'headers' in
the hobbies section.